requirements.txt
python==3.10.6
python-dotenv==1.0.0
langchain==0.0.137
pinecone-client==2.2.1
flask
flask_cors
gunicorn
openai==0.27.8
langchain_openai